fix: get info service type error
