+++ internal/form/answer/handler.go
@@ -20,6 +20,7 @@ import (
 	"github.com/NYCU-SDC/summer/pkg/problem"
 	"github.com/go-playground/validator/v10"
 	"github.com/google/uuid"
+	"github.com/jackc/pgx/v5/pgtype"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
@@ -80,7 +81,7 @@ type ResponseStore interface {
 	// GetSubmittedBy returns the user ID who submitted the response with the given ID.
 	// Used for ownership checks without creating an import cycle with the response package.
 	GetSubmittedBy(ctx context.Context, id uuid.UUID) (uuid.UUID, error)
-	GetEditInfo(ctx context.Context, id uuid.UUID) (GetEditInfoRow, error)
+	GetEditInfo(ctx context.Context, id uuid.UUID) (string, bool, pgtype.Timestamptz, error)
 }
 
 // OAuthProvider is the interface needed to initiate and complete an OAuth flow.
@@ -271,19 +272,19 @@ func (h *Handler) UpdateFormResponse(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Check edit info
-	editInfo, err := h.responseStore.GetEditInfo(traceCtx, responseID)
+	progress, allowEditResponse, deadline, err := h.responseStore.GetEditInfo(traceCtx, responseID)
 	if err != nil {
 		h.problemWriter.WriteError(traceCtx, w, err, logger)
 		return
 	}
 
-	if editInfo.Progress == "submitted" {
-		if !editInfo.AllowEditResponse {
+	if progress == "submitted" {
+		if !allowEditResponse {
 			h.problemWriter.WriteError(traceCtx, w, internal.ErrResponseEditNotAllowed, logger)
 			return
 		}
 
-		if editInfo.Deadline.Valid && time.Now().After(editInfo.Deadline.Time) {
+		if deadline.Valid && time.Now().After(deadline.Time) {
 			h.problemWriter.WriteError(traceCtx, w, internal.ErrFormDeadlineExceeded, logger)
 			return
 		}
