+++ internal/form/answer/queries.sql.go
@@ -143,29 +143,6 @@ func (q *Queries) GetByResponseIDAndQuestionID(ctx context.Context, arg GetByRes
 	return i, err
 }
 
-const getEditInfo = `-- name: GetEditInfo :one
-SELECT
-    r.progress,
-    f.allow_edit_response,
-    f.deadline
-FROM form_responses r
-         JOIN forms f ON f.id = r.form_id
-WHERE r.id = $1
-`
-
-type GetEditInfoRow struct {
-	Progress          ResponseProgress
-	AllowEditResponse bool
-	Deadline          pgtype.Timestamptz
-}
-
-func (q *Queries) GetEditInfo(ctx context.Context, id uuid.UUID) (GetEditInfoRow, error) {
-	row := q.db.QueryRow(ctx, getEditInfo, id)
-	var i GetEditInfoRow
-	err := row.Scan(&i.Progress, &i.AllowEditResponse, &i.Deadline)
-	return i, err
-}
-
 const getIDByResponseIDAndQuestionID = `-- name: GetIDByResponseIDAndQuestionID :one
 SELECT id FROM answers
 WHERE response_id = $1 AND question_id = $2
